Update to pedantic 1.8.0, also no implicit casts (#50)
* Update to pedantic 1.8.0, also no implicit casts

* format file

* Update min SDK

* Add a changelog entry

+++ test/int32_test.dart
@@ -115,15 +115,15 @@ void main() {
 
     test("remainder", () {
       expect(Int32(0x12345678).remainder(Int32(0x22)),
-          Int32(0x12345678.remainder(0x22)));
+          Int32(0x12345678.remainder(0x22) as int));
       expect(Int32(0x12345678).remainder(Int32(-0x22)),
-          Int32(0x12345678.remainder(-0x22)));
+          Int32(0x12345678.remainder(-0x22) as int));
       expect(Int32(-0x12345678).remainder(Int32(-0x22)),
-          Int32(-0x12345678.remainder(-0x22)));
+          Int32(-0x12345678.remainder(-0x22) as int));
       expect(Int32(-0x12345678).remainder(Int32(0x22)),
-          Int32(-0x12345678.remainder(0x22)));
+          Int32(-0x12345678.remainder(0x22) as int));
       expect(Int32(0x12345678).remainder(Int64(0x22)),
-          Int32(0x12345678.remainder(0x22)));
+          Int32(0x12345678.remainder(0x22) as int));
       expect(() => Int32(17).remainder(null), throwsArgumentError);
     });
